+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next
 
+* Item popup shows the ammo type of D2 weapons.
+* New is:primary, is:special, and is:heavy search terms for ammo types.
+* Add is:tracerifle and is:linearfusionrifle searches.
+* Added Korean as a language option.
 * We have a new Shop selling enamel pins and T-shirts.
 
 # 4.67.0 (2018-08-26)